Add is_bad to App Obj
Prints out the total number of apps and the number of apps with
something wrong. When is_bad is True, this is typically due to something
wrong from parsing values.yaml. Sometimes the app's repo is not ibmcom
but actually the name of the app. Starting to reduce size of black_list

Signed-off-by: Mick Tarsel <mtarsel@gmail.com>