Ensure that only 1 values.yaml is extracted
in this commit there is a lot of print outs. I've added a check to make
sure we get the values.yaml and the Chart.yaml from the top directory of
the Application. This is done by checking how many slashes are in the
file path of the YAML file. Also, ibm-microclimate is being testedit()

Signed-off-by: Mick Tarsel <mtarsel@gmail.com>