ibm-ace-server-dev prints out correctly
this was the last member of the black_list. this app is using the same
tag for all the images. the images and repos are stored together in a
dict which is the only member of a list. This commit contains a missed
called to black_list so will fail but app prints out correctly

Signed-off-by: Mick Tarsel <mtarsel@gmail.com>