Merge from python3 fix, product name (#4)
* Fixes #1 Fixes #3

python3 and travis badge

Signed-off-by: Ethan Hansen <1ethanhansen@gmail.com>

* Obtain Product name for each App

includes 2 more imports to utilize Github API to grab the first line of
the READMEs for each app in the IBM/charts/stable dir.

Prints out app name and product name and exits. Does nothing with
product name right now

Signed-off-by: Mick Tarsel <mtarsel@gmail.com>

* Use index.yaml or download latest from IBM/charts

cleaning up argparse

Signed-off-by: Mick Tarsel <mtarsel@gmail.com>

* user.yaml is NEEDED TO RUN, args clean up

In order to handle different repos and now the github token, there is a
new file needed to run get-image-info.py This file is a yaml file that
must contain user,password,token base64encded. This removes the need for
generated_input.yaml unless we run --debug mode.

Signed-off-by: Mick Tarsel <mtarsel@gmail.com>

* updated install and execute steps

user.yaml is required and moving to python3

* General code cleanup- remove productname() calls

Cleaning up the output for master branch. Some comments and
re-structuring but nothing major

Signed-off-by: Mick Tarsel <mtarsel@gmail.com>

* TODO: a single testit() not working

This helpful function is a copy of older source code. It _is_ a helpful
function for a single application but needs to be updated.

Signed-off-by: Mick Tarsel <mtarsel@gmail.com>

* Output product_name

Only grabs the first line of the README from the app so it's not perfect
but generally works.

Signed-off-by: Mick Tarsel <mtarsel@gmail.com>

* Output now inclues Product Name

Signed-off-by: Mick Tarsel <mtarsel@gmail.com>


Signed-off-by: Ethan Hansen <1ethanhansen@gmail.com>