Fixes #20, #22, #24 - rm indexparser.py & tests.py
There is no longer a need for indexparser.py except for mkdir_p() which
has been moved to utils/setup_utils.py

About #7, tests.py no longer exists since the way to test a single App
is in dire need of update. This file was riddled with old, repetitive
code and therefore has been removed.

For #20, all of the functions in get_image_info.py have been moved to
utils/setup_utils.py or have become methods for the App and Image
objects. This commit will close this issue.

Issue #22 is closed. This project no longer requires a github token
since we are downloading raw files from github. The README has been
updated to reflect this.

Finally, #24 is closed here as well. There is a regex to get Product
name.

Signed-off-by: Mick Tarsel <mtarsel@gmail.com>